add some shit code

+++ hami-common/src/main/java/top/wang3/hami/common/dto/request/LikeItemParam.java
@@ -7,5 +7,5 @@ import lombok.Data;
 public class LikeItemParam {
 
     private int itemId;
-    private int itemType;
+    private byte itemType;
 }
